Average-Pathlength filtration-curves
In this script I tried to execute the average path length graph descriptor function on the unlabeled dataset e.g. MUTAG and got accuracy: 80.61 +- 0.75

The igraph library reads each file path to load the graphs into memory. From the graph objects, the labels of the graphs are taken out and kept in the y list. The average_path_length method of igraph is used to get the average path length for each graph. Following that, the average route lengths are binarized by comparing each value to the lengths and turning it to a value of 0 or 1. 
The binarized data are padded to have a constant length, which produces the filtering curves. The create_curves method returns the filtering curves and the labels (y).